Moved the FingerprintGenerator related tests to a new file
